Simulazione mia

+++ src/main/java/it/polito/tdp/rivers/model/Flow.java
@@ -4,10 +4,10 @@ import java.time.LocalDate;
 
 public class Flow {
 	private LocalDate day;
-	private double flow;
+	private float flow;
 	private River river;
 
-	public Flow(LocalDate day, double flow, River river) {
+	public Flow(LocalDate day, float flow, River river) {
 		this.day = day;
 		this.flow = flow;
 		this.river = river;
@@ -21,11 +21,11 @@ public class Flow {
 		this.day = day;
 	}
 
-	public double getFlow() {
+	public float getFlow() {
 		return flow;
 	}
 
-	public void setFlow(double flow) {
+	public void setFlow(float flow) {
 		this.flow = flow;
 	}
